requirements.txt
numpy
scipy
pillow
requests
scikit-learn
streamlit
